requirements.txt
allure-python-commons
allure-python-commons
allure-python-commons
pytest
requests
PyYAML
pytest-order
pytest-html
jsonschema
allure-pytest